requirements.txt
aio-pika==9.5.4
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.8.0
cached-property==2.0.1
certifi==2024.12.14
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
loguru==0.7.3
multidict==6.1.0
pamqp==3.3.0
pika==1.3.2
portkey-ai==1.10.1
propcache==0.2.1
pydantic==2.10.5
pydantic-core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
tenacity==9.0.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
